requirements.txt
scrapy==2.11.0
twisted==22.10.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
requests==2.31.0
tabulate==0.9.0
